requirements.txt
streamlit
pillow
requests
python-dotenv
pydantic>=2.0.0
tenacity>=8.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pdfplumber
reportlab
scikit-learn
pandas
numpy
matplotlib
huggingface_hub>=0.20.0
groq
google-generativeai>=0.8.0
pytesseract
redis>=5.0.0
google-cloud-bigquery>=3.17.0
google-auth>=2.28.0
google-cloud-logging>=3.10.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.19.0
google-cloud-aiplatform>=1.49.0
